Removed icky angular
