--- server/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.ts",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "tsx server.ts",
-    "swagger": "node ./swagger.ts"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "nodemon": "^3.1.7",
-    "swagger-autogen": "^2.23.7",
-    "ts-node": "^10.9.2"
-  },
-  "dependencies": {
-    "@types/swagger-ui-express": "^4.1.6",
-    "dotenv": "^16.4.5",
-    "express": "^4.21.1",
-    "mongoose": "^8.7.1",
-    "swagger-ui-express": "^5.0.1",
-    "tsx": "^4.19.1"
-  }
-}
